[vector] Merge hb-vector-buf.cc into hb-vector.cc
The decimal point helper was the only content in the separate
.cc file.  Move it into hb-vector.cc and remove the extra
compilation unit.
